% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_results_rownames.R
\name{add_gene_info}
\alias{add_gene_info}
\title{Add Gene Information to Results Table}
\usage{
add_gene_info(results_table, anno_df)
}
\arguments{
\item{results_table}{A data frame containing the results. It must have row names or a column named "uniprot_id".}

\item{anno_df}{A data frame containing annotation information. It must have columns "unprot_id", "gene_symbol", and "ensembl_gene_id".}
}
\value{
A data frame with added columns "gene_symbol" and "ensembl_gene_id".
}
\description{
This function adds gene symbol and Ensembl gene ID information to a results table based on UniProt IDs.
}
\examples{
results_table <- data.frame(value = c(1.2, 3.4, 5.6), row.names = c("P12345", "Q67890", "A11111"))
anno_df <- data.frame(unprot_id = c("P12345", "Q67890", "A11111"), gene_symbol = c("GeneA", "GeneB", "GeneC"), ensembl_gene_id = c("ENSG000001", "ENSG000002", "ENSG000003"))
add_gene_info(results_table, anno_df)

}
