% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_symbol_to_dds.r
\name{add_symbols}
\alias{add_symbols}
\title{Add Gene Symbols to DESeq2 Dataset}
\usage{
add_symbols(dds, annotation_dataset)
}
\arguments{
\item{dds}{A DESeq2 dataset object.}

\item{annotation_dataset}{A data frame containing gene annotations with at least two columns:
\code{gene_id} and \code{gene_name}. The \code{gene_id} column should match the row names of the DESeq2 dataset.}
}
\value{
A DESeq2 dataset object with an additional column \code{SYMBOL} in the row data, containing the gene symbols.
}
\description{
This function adds gene symbols to the row data of a DESeq2 dataset object.
}
\examples{
\dontrun{
dds <- DESeqDataSetFromMatrix(countData = count_matrix, colData = col_data, design = ~ condition)
annotation_dataset <- data.frame(gene_id = c("gene1", "gene2"), gene_name = c("Gene A", "Gene B"))
dds <- add_symbols(dds, annotation_dataset)
}
}
