% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_datasets.r
\name{annotation_datasets}
\alias{annotation_datasets}
\title{Annotate Datasets}
\usage{
annotation_datasets(dds)
}
\arguments{
\item{dds}{A DESeq2 dataset object.}
}
\value{
A list containing:
\item{anns}{A data frame with Ensembl gene IDs, external gene names, and descriptions.}
\item{anno_df}{A data frame with annotations retrieved using the \code{pcaExplorer} package.}
}
\description{
This function annotates datasets by retrieving gene annotations from the Ensembl database.
}
\examples{
\dontrun{
library(DESeq2)
dds <- makeExampleDESeqDataSet()
annotations <- annotation_datasets(dds)
}

}
