% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_annotations_proteomics.r
\name{create_annotations_proteomics}
\alias{create_annotations_proteomics}
\title{Create Annotations}
\usage{
create_annotations_proteomics(se)
}
\arguments{
\item{se}{A SummarizedExperiment object. The function will stop if the provided object is not of class 'SummarizedExperiment'.}
}
\value{
This function does not return a value. It saves the generated annotations and annotation dataframe as RDS files in the "analyses_data" directory.
Additionally, it assigns the annotations and annotation dataframe to the parent environment.
}
\description{
This function generates annotation datasets from a SummarizedExperiment object and saves them as RDS files.
}
\examples{
\dontrun{
se <- SummarizedExperiment(...) # Create or load a SummarizedExperiment object
create_annotations(se)
}

}
