% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_top_GO_graph.R
\name{.info_enrichrun}
\alias{.info_enrichrun}
\title{Printing some info before the enrichment runs}
\usage{
.info_enrichrun(n_de, n_de_selected, de_type, res_de = NULL)
}
\arguments{
\item{n_de}{Numeric, number of DE genes (in total)}

\item{n_de_selected}{Character vector, containing the selected DE genes}

\item{de_type}{Character string, specifying up/down/both direction of
DE regulation}

\item{res_de}{The \code{res_de} container as expected in most \code{mosdef} functions.}
}
\value{
Prints out an informative summary message.
}
\description{
Printing some info before the enrichment runs
}
\examples{
# .info_enrichrun(10, length(c("geneA", "geneB")), "up")
}
