% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_metrics.R
\name{geometricCV}
\alias{geometricCV}
\title{Calculate Geometric Coefficient of Variation (CV)}
\usage{
geometricCV(row_values)
}
\arguments{
\item{row_values}{A numeric vector of values for which the geometric CV is to be calculated.}
}
\value{
A numeric value representing the geometric CV.
}
\description{
This function calculates the geometric coefficient of variation (CV) for a given set of row values.
The calculation is based on log-transformed data.
}
\examples{
values <- c(1, 2, 3, 4, 5)
geometricCV(values)
}
