% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_interface.R
\name{get_dea_list_report}
\alias{get_dea_list_report}
\title{get_dea_list but does not fail when empty}
\usage{
get_dea_list_report(dde, ...)
}
\arguments{
\item{dde}{A DeeDeeExperiment object}
}
\value{
A list of the available tables in the fea slot
}
\description{
This is a wrapper to the get_dea_list function
}
