% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_interface.R
\name{get_fea_list_report}
\alias{get_fea_list_report}
\title{get_fea_list but does not fail when empty}
\usage{
get_fea_list_report(dde, ...)
}
\arguments{
\item{dde}{A DeeDeeExperiment object}
}
\value{
A list of the available tables in the fea slot
}
\description{
This is a wrapper to the get_fea_list function
}
