% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_missing_values.R
\name{handling_missing_values_and_merging}
\alias{handling_missing_values_and_merging}
\title{Handling Missing Values and Merging Technical Replicates}
\usage{
handling_missing_values_and_merging(se, handling_of_missing_values = "no")
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object containing the count matrix and experiment metadata.}

\item{handling_of_missing_values}{A character string specifying how to handle missing values.
Options are "no" (default), "delete", or "infer".}
}
\value{
A \code{SummarizedExperiment} object with merged technical replicates and handled missing values.
}
\description{
This function processes a \code{SummarizedExperiment} object by handling missing values and merging technical replicates.
}
\details{
The function performs the following steps:
\enumerate{
\item Extracts the count matrix and experiment metadata from the input \code{SummarizedExperiment} object.
\item Identifies unique biological samples by concatenating the group and biological replicate information.
\item Sorts technical replicates for each biological sample.
\item Merges the count matrix by averaging technical replicates and assigning NA when all technical replicates are missing.
\item Creates merged experiment metadata.
\item Creates a new \code{SummarizedExperiment} object with the merged data.
\item Handles missing values based on the specified method:
\itemize{
\item "delete": Removes rows with any missing values.
\item "infer": Infers missing values by replacing them with the minimum value in the row, if the proportion of missing values is less than 30\%.
}
}
}
\examples{
# Example usage:
# merged_se <- handling_missing_values_and_merging(se, handling_of_missing_values = "delete")

}
