% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_analyses.r
\name{load_analyses}
\alias{load_analyses}
\title{Load Analysis Results}
\usage{
load_analyses(params)
}
\arguments{
\item{params}{A list containing the parameters for loading the analysis results.
It must include the following element:
\itemize{
\item \code{analysis_name} (optional): A character string specifying the name of the experiment.
If provided, the function will attempt to load files from the corresponding folder
within the "analyses_data" directory. If not provided, the function will attempt to load
files from the current working directory.
}}
}
\value{
This function does not return a value. Instead, it assigns the loaded objects to the parent environment:
\itemize{
\item \code{myresuSet}: The result set object loaded from "myResuSet.RDS".
\item \code{dds}: The DESeq2 results object loaded from "dds_DE_results.RDS".
\item \code{dds_releveled}: The releveled DESeq2 results object loaded from "dds_DE_results_releveled.RDS".
\item \code{dds_interaction_term}: The interaction term DESeq2 results object loaded from "dds_DE_results_interaction_term.RDS".
}
}
\description{
This function loads analysis results from specified files based on the provided parameters.
}
\details{
The function first checks if the \code{analysis_name} parameter is provided and if the corresponding
directory exists within the "analyses_data" folder. If so, it attempts to load the specified RDS files
from that directory. If any file is not found, a message is displayed. If \code{analysis_name} is not
provided or the directory does not exist, the function attempts to load the files from the current working
directory.
}
\examples{
\dontrun{
params <- list(analysis_name = "experiment1")
load_analyses(params)
}

}
