% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_helper_functions.R
\name{load_helper_functions}
\alias{load_helper_functions}
\title{Load Helper Functions}
\usage{
load_helper_functions(folder = "helper_functions")
}
\arguments{
\item{folder}{A character string specifying the folder containing the R scripts to load. Default is "helper_functions".}
}
\value{
None. The function loads the R scripts into the current environment.
}
\description{
Loads all R scripts located in a specified folder on the user's computer.
Each script is sourced into the \emph{global environment}, making all functions
and objects defined in those scripts immediately available for use.
}
