% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_load_proteomics.R
\name{load_proteomics_data}
\alias{load_proteomics_data}
\title{Load Proteomics Data}
\usage{
load_proteomics_data(params)
}
\arguments{
\item{experiment_name}{A character string specifying the name of the experiment.}
}
\value{
This function does not return a value but loads the following objects into the parent environment:
\itemize{
\item \code{vdx}: The vdx object loaded from "vdx.rds".
\item \code{se}: The se object loaded from "se.rds".
\item \code{myresuSet} (optional): The results object loaded from "results_object.rds" if it exists.
}
}
\description{
This function loads proteomics data for a given experiment from specified RDS files.
}
\details{
The function checks for the existence of the specified directory and the required RDS files.
If any of the required files are missing, the function stops with an error message.
If the optional results object file is not found, a message is displayed.
}
\examples{
\dontrun{
load_proteomics_data("experiment_1")
}

}
