% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_results.R
\name{load_results}
\alias{load_results}
\title{Load Results Based on Workflow and Parameters}
\usage{
load_results(params)
}
\arguments{
\item{params}{A list containing the following elements:}
}
\description{
This function loads data objects from RDS files based on the specified workflow and parameters.
It validates the input parameters, determines the directory path, and loads the appropriate objects
into the parent environment. If \code{run_computations} is set to \code{TRUE}, the loading operations are skipped.
}
\details{
The function checks the existence of required RDS files in the analized_data subfolder of the analysis into the parent environment.
}
\examples{
\dontrun{
params <- list(
  run_computations = TRUE,
  analysis_name = "example_analysis"
)
load_results(params)
}

}
