% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_switch.r
\name{missing_switch}
\alias{missing_switch}
\title{Replace Missing Values in SummarizedExperiment Assay Data}
\usage{
missing_switch(se, force_type = "zero")
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object. The input object whose assay data
will be modified.}

\item{force_type}{A character string specifying the type of replacement to
perform. Default is \code{"zero"}. If \code{"zero"}, NA values in the assay data
will be replaced with zeros. If \code{"NA"}, zero values in the assay data will
be replaced with NAs.}
}
\value{
A \code{SummarizedExperiment} object with modified assay data.
}
\description{
This function replaces missing values (NA) in the assay data of a
SummarizedExperiment object with zeros or vice versa, based on the specified
\code{force_type}.
}
\examples{
# Assuming `se` is a SummarizedExperiment object with some NA values in the assay data
se <- missing_switch(se, force_type = "zero")

# Assuming `se` is a SummarizedExperiment object with some zero values in the assay data
se <- missing_switch(se, force_type = "NA")

}
