% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression.R
\name{plot_expression}
\alias{plot_expression}
\title{Plot Expression}
\usage{
plot_expression(protein, se_object, use_gene_name = FALSE, count_matrix = NULL)
}
\arguments{
\item{protein}{A character string specifying the protein or gene to plot.}

\item{se_object}{A SummarizedExperiment object containing the expression data.}

\item{use_gene_name}{A logical value indicating whether to use gene names (default is FALSE).}

\item{count_matrix}{An optional matrix of counts. If NULL, the count matrix from the SummarizedExperiment object is used.}
}
\value{
A ggplot object representing the expression plot.
}
\description{
This function generates a boxplot of logCPM values for each group, with the option to use gene names or protein IDs.
}
\examples{
# Assuming `se` is a SummarizedExperiment object and "GeneA" is a valid protein/gene ID
plot_expression("GeneA", se, use_gene_name = TRUE)
}
