% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_reads.R
\name{plot_n_reads}
\alias{plot_n_reads}
\title{Plot Number of Reads per Sample

This function creates a bar plot showing the total number of reads for each sample in a DESeq2 dataset.}
\usage{
plot_n_reads(dds)
}
\arguments{
\item{dds}{A DESeqDataSet object containing the count data.}
}
\value{
A ggplot2 object representing the bar plot of read counts per sample.
}
\description{
Plot Number of Reads per Sample

This function creates a bar plot showing the total number of reads for each sample in a DESeq2 dataset.
}
\examples{
\dontrun{
library(DESeq2)
dds <- DESeqDataSetFromMatrix(countData = count_matrix,
                              colData = col_data,
                              design = ~ condition)
plot <- plot_n_reads(dds)
print(plot)
}
}
