% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.r
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA}
\usage{
plot_pca(
  dge_list,
  color_encoding = "group",
  legend_title,
  matrix_type = "getCounts",
  polygon = FALSE,
  plot_title = "PCA plot"
)
}
\arguments{
\item{dge_list}{A DGEList object containing the count data and sample information.}

\item{color_encoding}{A character string specifying the column name in the sample information to be used for coloring the points. Default is "group".}

\item{legend_title}{A character string specifying the title of the legend.}

\item{matrix_type}{A character string specifying the type of matrix to be used for PCA. Options are "logCPM", "getOffset", "getCounts", and "log2". Default is "getCounts".}

\item{polygon}{A logical value indicating whether to draw polygons around groups. Default is FALSE.}

\item{plot_title}{A character string specifying the title of the plot. Default is "PCA plot".}
}
\value{
A ggplot2 object representing the PCA plot.
}
\description{
This function performs Principal Component Analysis (PCA) on a given DGEList object and plots the results using ggplot2.
}
\examples{
\dontrun{
dge_list <- DGEList(counts = counts_matrix, samples = sample_info)
plot_pca(dge_list, color_encoding = "group", legend_title = "Group", matrix_type = "logCPM", polygon = TRUE, plot_title = "PCA of Samples")
}
}
