% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_workspace_proteomics.r
\name{prepare_workspace_proteomics}
\alias{prepare_workspace_proteomics}
\title{Prepare Workspace}
\usage{
prepare_workspace_proteomics(params)
}
\arguments{
\item{params}{A list containing the following elements:
\itemize{
\item \code{sample_metadata_file}: Path to the experiment metadata file (xlsx format).
\item \code{load_gse}: Logical, whether to load the GSE object (default is FALSE).
\item \code{load_complete_dds}: Logical, whether to load the complete DDS object (default is FALSE).
\item \code{complete_dds_source}: Path to the complete DDS source file.
}}
}
\value{
This function does not return a value. It assigns the loaded data to the global environment.
}
\description{
This function prepares the workspace by loading various data files and assigning them to the global environment.
}
\details{
The function performs the following steps:
\itemize{
\item Checks if the experiment metadata file exists. If not, it stops with an error.
\item Loads the experiment metadata from the specified file and assigns it to the global environment as \code{experiment_metadata}.
\item If \code{load_gse} is TRUE, attempts to load the GSE object from "analyses_data/gse.RDS" and assigns it to the global environment as \code{gse}.
\item If \code{load_complete_dds} is TRUE, attempts to load the complete DDS object from the specified source file and assigns it to the global environment as \code{complete_dds}.
\item Checks if the annotation files ("anns.RDS" and "anno_df.RDS") exist in the "analyses_data" directory. If they exist, loads them and assigns them to the global environment as \code{anns} and \code{anno_df}, respectively.
}
}
\examples{
\dontrun{
params <- list(
  sample_metadata_file = "path/to/metadata.xlsx",
  load_gse = TRUE,
  load_complete_dds = TRUE,
  complete_dds_source = "path/to/complete_dds.RDS"
)
prepare_workspace(params)
}

}
