% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_results_rownames.R
\name{results_to_gene_id}
\alias{results_to_gene_id}
\title{Convert Results Table to Gene ID}
\usage{
results_to_gene_id(results_table, anno_df)
}
\arguments{
\item{results_table}{A data frame containing the results table. It should have a column named \code{uniprot_id}.}

\item{anno_df}{A data frame containing annotation information, which will be used to add gene information if necessary.}
}
\value{
A data frame with Ensembl gene IDs as row names and the \code{ensembl_gene_id} column removed.
}
\description{
This function converts the row names of a results table to Ensembl gene IDs.
If the \code{ensembl_gene_id} column is not present in the results table, it will be added using the \code{anno_df} annotation dataframe.
}
\details{
The function performs the following steps:
\itemize{
\item Checks if the \code{ensembl_gene_id} column is present in the results table. If not, it adds the gene information using \code{anno_df}.
\item Reorders the columns to place \code{ensembl_gene_id} first.
\item Issues warnings if there are NA or duplicated gene names.
\item Removes rows with NA or duplicated \code{ensembl_gene_id}.
\item Sets the row names of the results table to \code{ensembl_gene_id} and removes the \code{ensembl_gene_id} column.
}
}
\examples{
\dontrun{
results_table <- data.frame(uniprot_id = c("P12345", "Q67890"), some_value = c(1.2, 3.4))
anno_df <- data.frame(uniprot_id = c("P12345", "Q67890"), ensembl_gene_id = c("ENSG000001", "ENSG000002"))
results_to_gene_id(results_table, anno_df)
}

}
