% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_results_rownames.R
\name{results_to_prot_id}
\alias{results_to_prot_id}
\title{Convert Results Table Row Names to Protein IDs}
\usage{
results_to_prot_id(results_table, anno_df)
}
\arguments{
\item{results_table}{A data frame containing the results table with row names that need to be converted to protein IDs.}

\item{anno_df}{A data frame containing the annotation information, where one of the columns is \code{gene_symbol} and the row names are protein IDs.}
}
\value{
A data frame with the same content as \code{results_table}, but with row names replaced by protein IDs from \code{anno_df}. The column \code{prot_ID} is removed from the final output.
}
\description{
This function takes a results table and an annotation data frame, and replaces the row names of the results table with protein IDs from the annotation data frame.
}
\examples{
\dontrun{
results_table <- data.frame(matrix(ncol = 3, nrow = 3))
rownames(results_table) <- c("gene1", "gene2", "gene3")
anno_df <- data.frame(gene_symbol = c("gene1", "gene2", "gene3"), row.names = c("prot1", "prot2", "prot3"))
results_to_prot_id(results_table, anno_df)
}
}
