% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resuset_to_gtl.r
\name{resuset_to_gtl}
\alias{resuset_to_gtl}
\title{Convert a result set to a GeneTonicList}
\usage{
resuset_to_gtl(myresuSet, result_name, dds, anno_df, enrich_algo = "not_given")
}
\arguments{
\item{myresuSet}{A list containing the result sets.}

\item{result_name}{A character string specifying the name of the result to be used.}

\item{dds}{A DESeqDataSet object.}

\item{anno_df}{A data frame containing annotation information.}

\item{enrich_algo}{A character string specifying the enrichment algorithm to be used.
Options are "not_given", "topGO_tbl", or "clupro_tbl". Default is "not_given".}
}
\value{
A GeneTonicList object containing the DESeqDataSet, DE results, enrichment results, and annotation data.
}
\description{
This function converts a result set to a GeneTonicList for different contrasts.
}
\examples{
# Example usage:
myresuSet <- list(
  contrast1 = list(
    res_DESeq = res_de_example,
    topGO_tbl = topgo_tbl_example,
    clupro_tbl = clupro_tbl_example
  )
)
result_name <- "contrast1"
dds <- dds_example
anno_df <- anno_df_example
gtl <- resuset_to_gtl(myresuSet, result_name, dds, anno_df, enrich_algo = "topGO_tbl")
}
