% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_top_GO_graph.R
\name{run_topGO_graph}
\alias{run_topGO_graph}
\title{Run topGO Graph}
\usage{
run_topGO_graph(
  de_container = NULL,
  res_de = NULL,
  de_genes = NULL,
  bg_genes = NULL,
  top_de = NULL,
  FDR_threshold = 0.05,
  min_counts = 0,
  ontology = "BP",
  annot = annFUN.org,
  mapping = "org.Mm.eg.db",
  gene_id = "symbol",
  full_names_in_rows = TRUE,
  add_gene_to_terms = TRUE,
  de_type = "up_and_down",
  topGO_method2 = "elim",
  do_padj = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{de_container}{A container object for differential expression analysis results. Default is NULL.}

\item{res_de}{A data frame containing the results of differential expression analysis. Default is NULL.}

\item{de_genes}{A vector of differentially expressed genes. Default is NULL.}

\item{bg_genes}{A vector of background genes. Default is NULL.}

\item{top_de}{A vector of top differentially expressed genes. Default is NULL.}

\item{FDR_threshold}{A numeric value specifying the false discovery rate threshold. Default is 0.05.}

\item{min_counts}{An integer specifying the minimum counts for filtering genes. Default is 0.}

\item{ontology}{A character string specifying the ontology to use ("BP" for Biological Process or "MF" for Molecular Function). Default is "BP".}

\item{annot}{A function for creating the topGO object. Default is annFUN.org.}

\item{mapping}{A character string specifying the annotation mapping to use. Default is "org.Mm.eg.db".}

\item{gene_id}{A character string specifying the type of gene identifier ("symbol" or "entrez"). Default is "symbol".}

\item{full_names_in_rows}{A logical value indicating whether to use full names in rows. Default is TRUE.}

\item{add_gene_to_terms}{A logical value indicating whether to add genes to terms. Default is TRUE.}

\item{de_type}{A character string specifying the type of differential expression ("up_and_down"). Default is "up_and_down".}

\item{topGO_method2}{A character string specifying the second method for topGO analysis. Default is "elim".}

\item{do_padj}{A logical value indicating whether to adjust p-values. Default is FALSE.}

\item{verbose}{A logical value indicating whether to print detailed messages. Default is TRUE.}
}
\value{
A topGO result object containing the enriched GO terms.
}
\description{
This function runs a topGO analysis to identify enriched GO terms in a set of differentially expressed genes.
}
\examples{
\dontrun{
result <- run_topGO_graph(de_container = my_de_container, res_de = my_res_de, de_genes = my_de_genes)
}

}
