% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_analyses_objects.r
\name{save_analysis_objects}
\alias{save_analysis_objects}
\title{Save Analysis Objects}
\usage{
save_analysis_objects(
  params,
  myresuSet,
  dds,
  dds_releveled = NULL,
  dds_interaction_term = NULL,
  gtl = NULL
)
}
\arguments{
\item{params}{A list containing parameters for the analysis. Must include \code{analysis_name} if saving to a specific directory.}

\item{myresuSet}{The main results set to be saved.}

\item{dds}{The DESeq2 results object to be saved.}

\item{dds_releveled}{Optional. A DESeq2 results object with releveled factors to be saved.}

\item{dds_interaction_term}{Optional. A DESeq2 results object with interaction terms to be saved.}

\item{gtl}{Optional. A gene-to-locus mapping object to be saved.}
}
\value{
None. This function is called for its side effects.
}
\description{
This function saves various analysis objects to disk. If an experiment name is provided in the \code{params} list,
the objects are saved in a directory named after the experiment within the \code{analyses_data} directory.
Otherwise, the objects are saved in the current working directory.
}
\examples{
\dontrun{
params <- list(analysis_name = "experiment_1")
save_analysis_objects(params, myresuSet, dds, dds_releveled, dds_interaction_term, gtl)
}
}
