% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_load_proteomics.R
\name{save_proteomics_data}
\alias{save_proteomics_data}
\title{Save Proteomics Data}
\usage{
save_proteomics_data(
  vdx,
  se,
  experiment_name,
  params = list(),
  results_object = NULL
)
}
\arguments{
\item{vdx}{A data object to be saved as an RDS file.}

\item{se}{A data object to be saved as an RDS file.}

\item{experiment_name}{A character string specifying the name of the experiment. This will be used to create a directory under "Analyses_data".}

\item{results_object}{An optional data object to be saved as an RDS file. Default is NULL.}
}
\value{
This function does not return a value. It is called for its side effects of saving data to disk.
}
\description{
This function saves proteomics data objects to specified directories.
}
\details{
The function creates a directory named after the \code{experiment_name} under "Analyses_data" if it does not already exist. It then saves the \code{vdx} and \code{se} objects as RDS files in this directory. If a \code{results_object} is provided, it is also saved as an RDS file in the same directory.
}
\examples{
\dontrun{
vdx <- data.frame()  # Example data
se <- data.frame()   # Example data
results_object <- list()  # Example data
save_proteomics_data(vdx, se, "experiment_1", results_object)
}

}
