% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_se.r
\name{save_se}
\alias{save_se}
\title{Save SummarizedExperiment Object}
\usage{
save_se(se, params)
}
\arguments{
\item{se}{A SummarizedExperiment object to be saved.}

\item{params}{A list of parameters, which should include an element named "analysis_name"
that specifies the name of the experiment. This name will be used to construct
the file path where the SE object will be saved.}
}
\value{
None. This function is used for its side effect of saving the SE object to a file.
}
\description{
This function saves a SummarizedExperiment (SE) object to a specified file path.
}
\examples{
\dontrun{
se <- SummarizedExperiment::SummarizedExperiment()
params <- list(analysis_name = "my_experiment")
save_se(se, params)
}

}
