% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deedee_interface.R
\name{show_interactive_table}
\alias{show_interactive_table}
\title{Display an interactive table with a custom caption}
\usage{
show_interactive_table(table, title, knitting = FALSE)
}
\arguments{
\item{table}{A data frame to display.}

\item{title}{A character string for the table caption.}

\item{knitting}{Logical; if \code{TRUE}, formats the table for knitting (e.g., RMarkdown).}
}
\value{
Displays the interactive table or prints a message if the table is empty.
}
\description{
This function shows a data frame as an interactive table using the DT package.
If the table is empty, a message is printed instead. When knitting a report,
the table is formatted for inclusion in the output document.
}
