% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_metrics.R
\name{standardCV}
\alias{standardCV}
\title{Calculate Coefficient of Variation (CV) for Proteomics Quality Control}
\usage{
standardCV(row_values)
}
\arguments{
\item{row_values}{A numeric vector of values for which the CV is to be calculated.}
}
\value{
The CV value as a numeric value.
}
\description{
These functions calculate the Coefficient of Variation (CV) for proteomics quality control.
}
\details{
\code{standardCV} calculates the CV using raw counts.
\code{geometricCV} calculates the CV using log-transformed data.
}
\examples{
# Example usage:
raw_counts <- c(10, 12, 15, 20, 25)
log_counts <- log(raw_counts)

standardCV(raw_counts)
geometricCV(log_counts)

}
