% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_annotations.r
\name{create_annotations}
\alias{create_annotations}
\title{Create Annotations}
\usage{
create_annotations(dds)
}
\arguments{
\item{dds}{A DESeqDataSet object. The function will stop if the provided object is not of class 'DESeqDataSet'.}
}
\value{
This function does not return a value. It saves the generated annotations and annotation dataframe as RDS files in the "analyses_data" directory.
Additionally, it assigns the annotations and annotation dataframe to the parent environment.
}
\description{
This function generates annotation datasets from a DESeqDataSet object and saves them as RDS files.
}
\examples{
\dontrun{
dds <- DESeqDataSet(...) # Create or load a DESeqDataSet object
create_annotations(dds)
}

}
