% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_results}
\alias{save_results}
\title{Save Multiple Objects Based on Conditions}
\usage{
save_results(params, objects)
}
\arguments{
\item{params}{A list containing the following elements:
\describe{
\item{\code{workflow}}{A string specifying the workflow type. Valid values are "se", "dds", "se_vdx", "se_dds".}
\item{\code{run_computations}}{A logical value indicating whether computations are run.}
\item{\code{analysis_folder}}{A string specifying the folder where analysis results should be saved.}
\item{\code{analysis_name}}{A string specifying the name of the analysis.}
\item{\code{save_results}}{A logical value indicating whether to save the results object.}
}}

\item{objects}{A named list of objects to save. Valid object names are:
\describe{
\item{\code{se}}{SummarizedExperiment object.}
\item{\code{dds}}{DESeqDataSet object.}
\item{\code{vdx}}{VDX object.}
\item{\code{results}}{Results object.}
}}
}
\value{
A message indicating that the objects were saved successfully.
}
\description{
This function saves multiple R objects to disk based on specified conditions and parameters.
It validates the input parameters and objects, creates necessary directories, and saves the objects
in \code{.rds} format according to the workflow and other conditions provided.
}
