% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_metadata.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Validate Metadata Against SummarizedExperiment Object}
\usage{
validate_metadata(metadata_df, se)
}
\arguments{
\item{metadata_df}{A \code{data.frame} containing metadata. The row names of this
data frame should correspond to the row names of the \code{rowData} in the
\code{SummarizedExperiment} object.}

\item{se}{A \code{SummarizedExperiment} object. The function validates that the
row names of the \code{metadata_df} match the row names of the \code{rowData} in this object.}
}
\value{
Returns \code{TRUE} if the row names match. If the row names do not match,
the function throws an error.
}
\description{
This function checks whether the row names of a metadata data frame match
the row names of the \code{rowData} in a \code{SummarizedExperiment} object.
}
\examples{
# Example usage:
# metadata_df <- data.frame(sample_id = c("sample1", "sample2"), row.names = c("sample1", "sample2"))
# se <- SummarizedExperiment::SummarizedExperiment(rowData = DataFrame(row.names = c("sample1", "sample2")))
# validate_metadata(metadata_df, se)

}
